+++ src/mock-constants.js
@@ -1,0 +1,48 @@
+"use strict";
+
+const TITLES = [
+  `Продам книги Стивена Кинга`,
+  `Продам новую приставку Sony Playstation 5`,
+  `Продам отличную подборку фильмов на VHS`,
+  `Куплю антиквариат`,
+  `Куплю породистого кота`,
+];
+
+const SENTENCES = [
+  `Товар в отличном состоянии.`,
+  `Пользовались бережно и только по большим праздникам.`,
+  `Продаю с болью в сердце...`,
+  `Бонусом отдам все аксессуары.`,
+  `Даю недельную гарантию.`,
+  `Если товар не понравится — верну всё до последней копейки.`,
+  `Это настоящая находка для коллекционера!`,
+  `Если найдёте дешевле — сброшу цену.`,
+  `Таких предложений больше нет!`,
+  `При покупке с меня бесплатная доставка в черте города.`,
+];
+
+const CATEGORIES = [`Книги`, `Разное`, `Посуда`, `Игры`, `Животные`, `Журналы`];
+
+const OfferType = {
+  OFFER: `offer`,
+  SALE: `sale`,
+};
+
+const SumRestrict = {
+  MIN: 1000,
+  MAX: 100000,
+};
+
+const PictureRestrict = {
+  MIN: 1,
+  MAX: 16,
+};
+
+module.exports = {
+  TITLES,
+  SENTENCES,
+  CATEGORIES,
+  OfferType,
+  SumRestrict,
+  PictureRestrict
+};
