+++ src/service/cli/generate.js
@@ -1,53 +1,18 @@
 "use strict";
 const fs = require(`fs`);
-// const fs = require(`fs`).promises;
-// const fs = require("fs/promises");
-
-const chalk = require("chalk");
-const { ExitCode } = require(`../../constants`);
-const { getRandomInt, shuffle } = require(`../../utils`);
+const chalk = require(`chalk`);
+const util = require(`util`);
+const {ExitCode} = require(`../../constants`);
+const {getRandomInt, shuffle} = require(`../../utils`);
+const {TITLES,
+  SENTENCES,
+  CATEGORIES,
+  OfferType,
+  SumRestrict,
+  PictureRestrict} = require(`../../mock-constants`);
 
 const DEFAULT_COUNT = 1;
 const FILE_NAME = `mocks.json`;
-
-const TITLES = [
-  `Продам книги Стивена Кинга`,
-  `Продам новую приставку Sony Playstation 5`,
-  `Продам отличную подборку фильмов на VHS`,
-  `Куплю антиквариат`,
-  `Куплю породистого кота`,
-];
-
-const SENTENCES = [
-  `Товар в отличном состоянии.`,
-  `Пользовались бережно и только по большим праздникам.`,
-  `Продаю с болью в сердце...`,
-  `Бонусом отдам все аксессуары.`,
-  `Даю недельную гарантию.`,
-  `Если товар не понравится — верну всё до последней копейки.`,
-  `Это настоящая находка для коллекционера!`,
-  `Если найдёте дешевле — сброшу цену.`,
-  `Таких предложений больше нет!`,
-  `При покупке с меня бесплатная доставка в черте города.`,
-];
-
-const CATEGORIES = [`Книги`, `Разное`, `Посуда`, `Игры`, `Животные`, `Журналы`];
-
-const OfferType = {
-  OFFER: `offer`,
-  SALE: `sale`,
-};
-
-const SumRestrict = {
-  MIN: 1000,
-  MAX: 100000,
-};
-
-const PictureRestrict = {
-  MIN: 1,
-  MAX: 16,
-};
-
 const MAX_OFFERS = 1000;
 
 const getPictureFileName = (number) =>
@@ -59,7 +24,7 @@ const generateOffers = (count) =>
     .map(() => ({
       title: TITLES[getRandomInt(0, TITLES.length - 1)],
       picture: getPictureFileName(
-        getRandomInt(PictureRestrict.MIN, PictureRestrict.MAX)
+          getRandomInt(PictureRestrict.MIN, PictureRestrict.MAX)
       ),
       description: shuffle(SENTENCES).slice(1, 5).join(` `),
       type: Object.keys(OfferType)[
@@ -71,7 +36,7 @@ const generateOffers = (count) =>
 
 module.exports = {
   name: `--generate`,
-  run(args) {
+  async run(args) {
     const [count] = args;
     const countOffer = Number.parseInt(count, 10) || DEFAULT_COUNT;
     if (countOffer > MAX_OFFERS) {
@@ -80,14 +45,13 @@ module.exports = {
     }
     const content = JSON.stringify(generateOffers(countOffer));
 
-    fs.writeFile(FILE_NAME, content, (err) => {
-      if (err) {
-        console.error(chalk.red(`Can't write data to file...`));
-        process.exit(ExitCode.error);
-      }
-
+    try {
+      await util.promisify(fs.writeFile)(FILE_NAME, content);
       console.info(chalk.green(`Operation success. File created.`));
       process.exit(ExitCode.success);
-    });
+    } catch (err) {
+      console.error(chalk.red(`Can't write data to file...`));
+      process.exit(ExitCode.error);
+    }
   },
 };
