+++ src/service/cli/version.js
@@ -1,5 +1,5 @@
 "use strict";
-const chalk = require("chalk");
+const chalk = require(`chalk`);
 const packageJsonFile = require(`../../../package.json`);
 const { ExitCode } = require(`../../constants`);
 
@@ -7,7 +7,7 @@ module.exports = {
   name: `--version`,
   run() {
     const version = packageJsonFile.version;
-    console.info(chalk.magenta(version));
+    console.info(chalk.blue(version));
     process.exit(ExitCode.success);
   },
 };
